requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
chardet==4.0.0
distro==1.8.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==2.10
numpy==1.26.1
openai==1.2.3
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
requests==2.25.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
